reduce to object-level barrier
