+++ rts/motoko-rts/src/gc/mark_compact.rs
@@ -34,20 +34,14 @@ unsafe fn schedule_compacting_gc<M: Memory>(mem: &mut M) {
 
 #[ic_mem_fn(ic_only)]
 unsafe fn compacting_gc<M: Memory>(mem: &mut M) {
-    use super::write_barrier::{init_write_barrier, REMEMBERED_SET};
     use crate::memory::ic;
 
-    let using_write_barrier = REMEMBERED_SET.is_some();
-    if using_write_barrier {
-        #[cfg(debug_assertions)]
-        super::write_barrier::sanity_checks::verify_snapshot(
-            ic::get_aligned_heap_base() as usize,
-            ic::HP as usize,
-            ic::get_static_roots(),
-        );
-        // remembered set is collected by GC
-        REMEMBERED_SET = None;
-    }
+    #[cfg(debug_assertions)]
+    super::write_barrier::sanity_checks::verify_snapshot(
+        ic::get_aligned_heap_base() as usize,
+        ic::HP as usize,
+        ic::get_static_roots(),
+    );
 
     compacting_gc_internal(
         mem,
@@ -66,11 +60,8 @@ unsafe fn compacting_gc<M: Memory>(mem: &mut M) {
 
     ic::LAST_HP = ic::HP;
 
-    if using_write_barrier {
-        #[cfg(debug_assertions)]
-        super::write_barrier::sanity_checks::take_snapshot(mem, ic::HP as usize);
-        init_write_barrier(mem);
-    }
+    #[cfg(debug_assertions)]
+    super::write_barrier::sanity_checks::take_snapshot(mem, ic::HP as usize);
 }
 
 pub unsafe fn compacting_gc_internal<
