+++ rts/motoko-rts/src/text_iter.rs
@@ -30,17 +30,9 @@ unsafe fn find_leaf<M: Memory>(mem: &mut M, mut text: Value, todo: *mut Value) -
         // Add right node to TODOs
         let new_todo = alloc_array(mem, 2);
         let new_todo_array = new_todo.as_array();
-        write_barrier(
-            mem,
-            new_todo,
-            new_todo_array.payload_addr().add(TODO_TEXT_IDX as usize) as u32,
-        );
+        write_barrier(mem, new_todo);
         new_todo_array.set(TODO_TEXT_IDX, (*concat).text2);
-        write_barrier(
-            mem,
-            new_todo,
-            new_todo_array.payload_addr().add(TODO_LINK_IDX as usize) as u32,
-        );
+        write_barrier(mem, new_todo);
         new_todo_array.set(TODO_LINK_IDX, *todo);
         *todo = new_todo;
 
@@ -70,11 +62,7 @@ pub unsafe fn text_iter<M: Memory>(mem: &mut M, text: Value) -> Value {
     array.set(ITER_POS_IDX, Value::from_scalar(0));
 
     // Initialize blob field
-    write_barrier(
-        mem,
-        iter,
-        array.payload_addr().add(ITER_BLOB_IDX as usize) as u32,
-    );
+    write_barrier(mem, iter);
     array.set(ITER_BLOB_IDX, find_leaf(mem, text, todo_addr as *mut _));
 
     iter
@@ -121,20 +109,12 @@ pub unsafe fn text_iter_next<M: Memory>(mem: &mut M, iter: Value) -> u32 {
             // allocation)
             let concat = text.as_concat();
 
-            write_barrier(
-                mem,
-                todo,
-                todo_array.payload_addr().add(TODO_TEXT_IDX as usize) as u32,
-            );
+            write_barrier(mem, todo);
             todo_array.set(TODO_TEXT_IDX, (*concat).text2);
             iter_array.set(ITER_POS_IDX, Value::from_scalar(0));
             let todo_addr = iter_array.payload_addr().add(ITER_TODO_IDX as usize);
 
-            write_barrier(
-                mem,
-                iter,
-                iter_array.payload_addr().add(ITER_BLOB_IDX as usize) as u32,
-            );
+            write_barrier(mem, iter);
             iter_array.set(ITER_BLOB_IDX, find_leaf(mem, (*concat).text1, todo_addr));
 
             text_iter_next(mem, iter)
@@ -142,19 +122,11 @@ pub unsafe fn text_iter_next<M: Memory>(mem: &mut M, iter: Value) -> u32 {
             // Otherwise remove the entry from the chain
             debug_assert_eq!(text.tag(), TAG_BLOB);
 
-            write_barrier(
-                mem,
-                iter,
-                iter_array.payload_addr().add(ITER_BLOB_IDX as usize) as u32,
-            );
+            write_barrier(mem, iter);
             iter_array.set(ITER_BLOB_IDX, text);
             iter_array.set(ITER_POS_IDX, Value::from_scalar(0));
 
-            write_barrier(
-                mem,
-                iter,
-                iter_array.payload_addr().add(TODO_LINK_IDX as usize) as u32,
-            );
+            write_barrier(mem, iter);
             iter_array.set(ITER_TODO_IDX, todo_array.get(TODO_LINK_IDX));
 
             text_iter_next(mem, iter)
