+++ src/codegen/compile.ml
@@ -934,7 +934,7 @@ module RTS = struct
     E.add_func_import env "rts" "stream_reserve" [I32Type; I32Type] [I32Type];
     E.add_func_import env "rts" "stream_stable_dest" [I32Type; I64Type; I64Type] [];
     E.add_func_import env "rts" "init_write_barrier" [] [];
-    E.add_func_import env "rts" "write_barrier" [I32Type; I32Type] [];
+    E.add_func_import env "rts" "write_barrier" [I32Type] [];
     ()
 
 end (* RTS *)
@@ -6775,12 +6775,9 @@ module Var = struct
       G.i (LocalSet (nr i))
 
     | Some (HeapInd i) ->
-      (if !Flags.write_barrier (* TODO: Optimize by determining whether the field type allows potential pointers *)
+      (if !Flags.write_barrier (* TODO: Optimize if the field does certainly not store a pointer *)
        then
-        G.i (LocalGet (nr i)) ^^ (* write barrier: object argument *)
         G.i (LocalGet (nr i)) ^^
-        compile_add_const ptr_unskew ^^
-        compile_add_const (Int32.mul MutBox.field Heap.word_size) ^^ (* write barrier: location argument *)
         E.call_import env "rts" "write_barrier"
        else G.nop
       ) ^^
@@ -6789,12 +6786,9 @@ module Var = struct
       Heap.store_field MutBox.field
 
     | Some (HeapStatic ptr) ->
-      (if !Flags.write_barrier (* TODO: Optimize by determining whether the field type allows potential pointers *)
+      (if !Flags.write_barrier (* TODO: Optimize if the field does certainly not store a pointer *)
        then 
-        compile_unboxed_const ptr ^^ (* write barrier: object argument *)
         compile_unboxed_const ptr ^^
-        compile_add_const ptr_unskew ^^
-        compile_add_const (Int32.mul MutBox.field Heap.word_size) ^^ (* write barrier: location argument *)
         E.call_import env "rts" "write_barrier"
        else G.nop
       ) ^^
@@ -8210,17 +8204,15 @@ let rec compile_lexp (env : E.t) ae lexp =
    ( 
     compile_exp_vanilla env ae e1 ^^ (* offset to array *)
     compile_exp_vanilla env ae e2 ^^ (* idx *)
+    Arr.idx_bigint env ^^
     (if !Flags.write_barrier && (potential_pointer (Arr.element_type env e1.note.Note.typ))
      then
-      Arr.idx_bigint env ^^
       set_field ^^
-      compile_exp_vanilla env ae e1 ^^ (* write barrier: object argument *)
-      get_field ^^
-      compile_add_const ptr_unskew ^^ (* write barrier: location argument *)
+      compile_exp_vanilla env ae e1 ^^
       E.call_import env "rts" "write_barrier" ^^
       get_field
      else
-      Arr.idx_bigint env
+      G.nop
      ),
      SR.Vanilla,
      store_ptr
@@ -8230,10 +8222,10 @@ let rec compile_lexp (env : E.t) ae lexp =
     
    ( 
      compile_exp_vanilla env ae e ^^
+     Object.idx env e.note.Note.typ n ^^
      (if !Flags.write_barrier && (potential_pointer (Object.field_type env e.note.Note.typ n))
       then  
         (* Only real objects have mutable fields, no need to branch on the tag *)
-        Object.idx env e.note.Note.typ n ^^
         set_field ^^
         (if Object.is_mut_field env e.note.Note.typ n
          then 
@@ -8244,12 +8236,10 @@ let rec compile_lexp (env : E.t) ae lexp =
           (* inlined field *)
           compile_exp_vanilla env ae e
         ) ^^ (* write barrier: object argument *)
-        get_field ^^
-        compile_add_const ptr_unskew ^^ (* write barrier: location argument *)
         E.call_import env "rts" "write_barrier" ^^
         get_field
       else
-        Object.idx env e.note.Note.typ n
+        G.nop
      ),
      SR.Vanilla,
      store_ptr
